feat(email): add scheduler module for campaigns
