+++ packages/email/src/index.ts
@@ -4,3 +4,9 @@ export type { AbandonedCart } from "./abandonedCart";
 export { recoverAbandonedCarts } from "./abandonedCart";
 export { sendEmail } from "./sendEmail";
 export { resolveSegment } from "./segments";
+export {
+  createCampaign,
+  listCampaigns,
+  sendDueCampaigns,
+} from "./scheduler";
+export type { Campaign } from "./scheduler";
